Update tests for new classifications
