+++ .gitignore
@@ -1,0 +1,3 @@
+.pytest_cache
+__pycache__
+*.pyc
